Refactor ArtistDetails and extract reusable components
