+++ src/components/Loading.tsx
@@ -1,0 +1,14 @@
+import { Box, CircularProgress } from "@mui/material";
+
+const Loading = () => (
+  <Box
+    display="flex"
+    justifyContent="center"
+    alignItems="center"
+    minHeight={200}
+  >
+    <CircularProgress />
+  </Box>
+);
+
+export default Loading;
